requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
craft-text-detector==0.4.3
filelock==3.8.0
gdown==4.5.3
idna==3.4
kafka-python==2.0.2
numpy==1.23.4
opencv-python==4.5.4.60
Pillow==9.2.0
PySocks==1.7.1
requests==2.28.1
scipy==1.9.3
six==1.16.0
soupsieve==2.3.2.post1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
